+++ sphinx/source/changelog.rst
@@ -25,6 +25,7 @@ Changelog
   more blocks on a tape)
 * Fixed how two base prefixes are handled when applied to an instruction with
   one operand
+* Fixed the parsing of a 48K SNA snapshot that has SP set to 0xFFFF
 
 10.1 (2026-08-14)
 -----------------